requirements.txt
Django==1.8.5
selenium
nose==1.3.0
djangorestframework
